+++ Assets/Scenes/Loading.unity
@@ -131,7 +131,7 @@ GameObject:
   - component: {fileID: 101353977}
   - component: {fileID: 101353976}
   m_Layer: 5
-  m_Name: Text (TMP)
+  m_Name: loadingText
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -153,7 +153,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: -347, y: -185}
+  m_AnchoredPosition: {x: -300.6, y: -171.2}
   m_SizeDelta: {x: 200, y: 50}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &101353976
