--- Assets/Script/Scence_Manager/ReturnMenu.cs
@@ -1,23 +1,0 @@
-using UnityEngine;
-using UnityEngine.UI;
-
-public class ReturnMenu : MonoBehaviour
-{
-    private void Start()
-    {
-        Button button = GetComponent<Button>();
-        if (button != null)
-        {
-            button.onClick.AddListener(OnImageClick);
-        }
-        else
-        {
-            Debug.LogWarning("Image không có Button component! Hãy thêm Button vào Image.");
-        }
-    }
-
-    private void OnImageClick()
-    {
-        Loader.Load(Loader.Scene.MenuScene);
-    }
-}
