+++ Assets/Script/Scence_Manager/StartGame.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 454ed4e99e9a35046979a45d9a7a6d93
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
