+++ Assets/Script/Scence_Manager/StartGame/StartGame.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+public class StartGame: MonoBehaviour
+{
+    public void OnStartGameButtonClicked()
+    {
+        Loader.Load(Loader.Scene.MainGameScene);
+    }
+}
