+++ Assets/Script/Scence_Manager/StartGame/startGameBtnController.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 6ca3c80744ad90d4d9cdeea57d33c387
