+++ Assets/Script/Scence_Manager/menu/ReturnMenu.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+public class ReturnMenu : MonoBehaviour
+{
+    public void OnMenuButtonClicked()
+    {
+        Loader.Load(Loader.Scene.MenuScene);
+    }
+}
